Reconciled units but the values still fall outside the default tolerance of assertAlmostEqual.
This problem is now a bit more difficult than I was expecting. The units problem is hacky, but more problematic is that the units I use within Electrode are SI, but the units I use in electrode_standard_data_generator.py are what I call TEC units. I need to understand why the values are so different from essentially the same calculation.