Draft test
This test tests the `tec.TECBase.calc_motive` method. I also modified
the `co` object so that the emitter and collector are a nonzero distance
apart. Finally, I abstracted the creation of a `tec.TECBase` object into
its own class so that I don't have to duplicate a bunch of `setUp`
methods.