Add check to protect against invalid input changes
I have some dummy input objects defined at the top of `test_TECBase.py`;
It is important that the position attribute of the `co` object is
greater than the position attribute of the `em` object because several
of the tests are based on that assumption. This check will cause many of
the tests to fail outright. The hope is that if someone changes these
dummy object attributes in a way that might break tests, this exception
will let them know what went wrong.